move ConvertableUnits to unitops
